+++ scripts/deploy/tearDown.sh
@@ -63,6 +63,6 @@ echo "Removing the Jars bucket and docker containers"
 "${THIS_DIR}/removeUploads.sh" "${INSTANCE_PROPERTIES}"
 
 echo "Removing generated files"
-rm -r "${GENERATED_DIR}"
+rm -r "${GENERATED_DIR:?}"/*
 
 echo "Successfully torn down"
